conversion warnings, sizes in custom dream classes

+++ DREAM/Examples/example_dream.cpp
@@ -17,7 +17,7 @@ int main(int argc, const char**){
  *
  */
 
-    srand(time(0));
+    srand((int) time(0));
 
     bool limit_examples = false;
     if (argc > 1){
@@ -43,13 +43,13 @@ int main(int argc, const char**){
         ~UnscaledPDF(){}
         int getNumDimensions() const{ return 1; }
         void evaluate(const std::vector<double> x, std::vector<double> &y, bool useLogForm){
-            int num_points = x.size();
-            if (y.size() < (size_t) num_points) y.resize(num_points);
-            for(int i=0; i<num_points; i++){ // set the pdf values
+            size_t num_points = x.size();
+            if (y.size() < num_points) y.resize(num_points);
+            for(size_t i=0; i<num_points; i++){ // set the pdf values
                 y[i] = -0.5 * x[i] * x[i];
             }
             if (!useLogForm){
-                for(int i=0; i<num_points; i++){ // take exponential (if not working with logForm)
+                for(size_t i=0; i<num_points; i++){ // take exponential (if not working with logForm)
                     y[i] = exp(y[i]);
                 }
             }
@@ -579,9 +579,9 @@ int main(int argc, const char**){
         int getNumDimensions() const{ return 2; }
         int getNumOutputs() const{ return N; }
         void evaluate(const std::vector<double> x, std::vector<double> &y) const{
-            int num_points = x.size() / 2;
-            if (y.size() < (size_t) (N * num_points)) y.resize(N * num_points);
-            for(int i=0; i<num_points; i++){
+            size_t num_points = x.size() / 2;
+            if (y.size() < N * num_points) y.resize(N * num_points);
+            for(size_t i=0; i<num_points; i++){
                 for(int j=0; j<N; j++){
                     y[i*N + j] = sin(x[2*i] * M_PI * (dt2 + j*dt) + x[2*i+1]);
                 }
@@ -704,9 +704,9 @@ int main(int argc, const char**){
         int getNumDimensions() const{ return 6; }
         int getNumOutputs() const{ return N; }
         void evaluate(const std::vector<double> x, std::vector<double> &y) const{
-            int num_points = x.size() / 2;
-            if (y.size() < (size_t) (N * num_points)) y.resize(N * num_points);
-            for(int i=0; i<num_points; i++){
+            size_t num_points = x.size() / 2;
+            if (y.size() < N * num_points) y.resize(N * num_points);
+            for(size_t i=0; i<num_points; i++){
                 for(int j=0; j<N; j++){
                     double t = snap[j];
                     y[i*N + j] = x[6*i];
